tickets/routes/__test__/new.test.ts
import request from "supertest";
import {app} from "../../src/app";
import {Ticket} from "../../src/models/ticket";

it('has a route handler listening to /api/tickets post post requests', async () => {
    const response = await request(app)
        .post('/api/tickets')
        .send({});
    expect(response.status).not.toEqual(404);
})

it('can be accessed if user is signed in', async () => {
    const response = await request(app)
        .post('/api/tickets')
        .set('Cookie', global.signin())
        .send({})
    expect(response.status).not.toEqual(401);
})

it('creates ticket with valid inputs', async () => {
    let tickets = await Ticket.find({})
    expect(tickets.length).toEqual(0)
    const title = 'abcde'
    await request(app)
        .post('/api/tickets')
        .set('Cookie', global.signin())
        .send({
            title,
            price: 20
        })
        .expect(201)
    tickets = await Ticket.find({})
    expect(tickets.length).toEqual(1)
    expect(tickets[0].price).toEqual(20)
    expect(tickets[0].title).toEqual(title)
})

it('returns error if invalid title is provided', async () => {

})

it('returns error if invalid price is provided', async () => {

})